+++ win32/3rdparty/pthreads/tests/valid1.c
@@ -94,7 +94,7 @@ main()
   washere = 0;
   assert(pthread_create(&t, NULL, func, NULL) == 0);
   assert(pthread_join(t, &result) == 0);
-  assert(result == 0);
+  assert((int)(size_t)result == 0);
   assert(washere == 1);
   sched_yield();
   assert(pthread_kill(t, 0) == ESRCH);
